Rename quiz creator service and add organization input for automations
